resolver: skip answers of unsupported IP versions
The previous code always started checking for an A record, falling back
to an AAAA if the first does not exists. Since most domains have an A
record - and if only for compatibility reasons - this one is also
returned on machines which do not even have an IPv4 connection,
resulting in connectivity issues.

This change will also take a look at the current machine's network
interfaces and only return A records if a local IPv4 address exists,
same for AAAA records and local IPv6 addresses.

This closes #1310.