add dns.REFUSED (#1458)
When a local DNS server does not allow recursion, it can return the error "Query Refused" as told in `**RFC1035 Section 4.1.1 RCODE 5**: Refused - The name server refuses to perform the specified operation for policy reasons.  For example, a name server may not wish toprovide the information to the particular requester, or a name server may not wish to perform a particular operation (e.g., zone transfer) for particular data.` So this is another case which the request shoud go to the next resolver.